requirements.txt
pandas
matplotlib
Pillow
numpy
seaborn
scikit-learn